Use updated protobufs and check friendCount for nil pointer

+++ handler/relation_handler/get_favorite_parties_by_user.go
@@ -18,7 +18,7 @@ func (h *relationGatewayHandler) GetFavoritePartiesByUser(c *fiber.Ctx) error {
 	limitStr := c.Query("limit")
 	limit, _ := strconv.ParseUint(limitStr, 10, 32)
 
-	fpRes, err := h.rc.GetFavoritePartiesByUser(c.Context(), &rg.GetFavoritePartiesByUserRequest{UserId: uId, NextPage: nextPage, Limit: uint32(limit)})
+	fpRes, err := h.rc.GetFavoritePartiesByUser(c.Context(), &rg.GetFavoritePartiesByUserRequest{UserId: uId, NextPage: nextPage, Limit: limit})
 	if err != nil {
 		return utils.ToHTTPError(err)
 	}
